Track Terragrunt version
With the current tag reflecting the terraform version, there have been
no updates to inclue newer versions of terragrunt.

This change clarifies where terragrunt vs terraform version is used.
It also uses a combined tag so that new builds will be generated when a
new version of Terraform or Terragrunt are released.